Refactored the SelectionUtil into separate algorithm types
