+++ PxWeb/Code/Api2/DataSelection/IDefaultSelectionAlgorithm.cs
@@ -1,0 +1,11 @@
+﻿using PCAxis.Paxiom;
+
+using PxWeb.Api2.Server.Models;
+
+namespace PxWeb.Code.Api2.DataSelection
+{
+    public interface IDefaultSelectionAlgorithm
+    {
+        VariablesSelection GetDefaultSelection(IPXModelBuilder builder);
+    }
+}
